requirements.txt
pyTelegramBotAPI
openai
python-docx
pymupdf